+++ src/background/utils/mod.rs
@@ -49,10 +49,6 @@ pub fn pcwstr(s: &str) -> windows::core::PCWSTR {
     windows::core::PCWSTR::from_raw(s.encode_utf16().chain(Some(0)).collect_vec().as_ptr())
 }
 
-pub fn sleep_millis(millis: u64) {
-    std::thread::sleep(Duration::from_millis(millis));
-}
-
 /// Resolve paths with folder ids in the form of "{GUID}\path\to\file"
 ///
 /// https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid